+++ src/main/java/refinedstorage/block/BlockConstructor.java
@@ -74,8 +74,16 @@ public class BlockConstructor extends BlockCable {
         return new TileConstructor();
     }
 
+    private boolean isInAABB(AxisAlignedBB aabb, float hitX, float hitY, float hitZ) {
+        return hitX >= aabb.minX && hitX <= aabb.maxX && hitY >= aabb.minY && hitY <= aabb.maxY && hitZ >= aabb.minZ && hitZ <= aabb.maxZ;
+    }
+
     @Override
     public boolean onBlockActivatedDefault(World world, BlockPos pos, IBlockState state, EntityPlayer player, EnumHand hand, ItemStack heldItem, EnumFacing side, float hitX, float hitY, float hitZ) {
+        if (hitCablePart(state, world, pos, hitX, hitY, hitZ)) {
+            return false;
+        }
+
         if (!world.isRemote) {
             player.openGui(RefinedStorage.INSTANCE, RefinedStorageGui.CONSTRUCTOR, world, pos.getX(), pos.getY(), pos.getZ());
         }