+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fixed not being able to move inventory items in Grid GUI's to hotbar via the number keys
 - Fixed Relays when being in "Ignore Redstone" mode using up energy
 - Improved collisions of Cable parts
+- You now have to click the actual cable part head in order to get the GUI open
 
 **Features**
 - Added German translation by ChillUpX